+++ bvproductsentiments/src/main/java/com/bazaarvoice/bvandroidsdk/ExpressionsSentimentResponse.java
@@ -1,0 +1,4 @@
+package com.bazaarvoice.bvandroidsdk;
+
+public class ExpressionsSentimentResponse extends Expressions{
+}