+++ bvproductsentiments/src/main/java/com/bazaarvoice/bvandroidsdk/ProductSentimentsFingerprintProvider.java
@@ -1,0 +1,15 @@
+package com.bazaarvoice.bvandroidsdk;
+
+import androidx.annotation.WorkerThread;
+
+public interface ProductSentimentsFingerprintProvider {
+  ProductSentimentsFingerprintProvider EMPTY = new ProductSentimentsFingerprintProvider() {
+    @Override
+    public String getFingerprint() {
+      return "";
+    }
+  };
+
+  @WorkerThread
+  String getFingerprint();
+}