+++ bvproductsentiments/src/main/java/com/bazaarvoice/bvandroidsdk/QuotesSentimentResponse.java
@@ -1,0 +1,4 @@
+package com.bazaarvoice.bvandroidsdk;
+
+public class QuotesSentimentResponse extends  Quotes{
+}