+++ bvproductsentiments/src/main/java/com/bazaarvoice/bvandroidsdk/SummarisedFeaturesQuotesResponse.java
@@ -1,0 +1,4 @@
+package com.bazaarvoice.bvandroidsdk;
+
+public class SummarisedFeaturesQuotesResponse extends Quotes{
+}